MVCtest2 5

+++ MVCtest2/src/main/java/ru/MVCtest/config/controllers/FirstController.java
@@ -39,7 +39,7 @@ public class FirstController {
                 result = a * b;
                 break;
             case "Division":
-                result = a / b;
+                result = a / (double)b;
                 break;
             }
         }
